requirements.txt
scikit-learn==0.24.2
joblib==1.0.1
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
Pillow==9.0.1
numpy==1.21.5